Solve day 3

+++ src/day01.rs
@@ -12,12 +12,12 @@ impl Solver for Day01 {
     type Output2 = usize;
 
     fn solve_part1(&self, input: &Self::Input) -> Self::Output1 {
-        // @Note: while the natural solution would be to use a HashMap,
+        // @Note: while the natural solution would be to use a HashSet,
         // using a Vec with less than 2048 positions will do just fine.
         let mut has_complement = [false; TARGET_SUM + 1];
 
         for &entry in input {
-            if has_complement[entry as usize] {
+            if has_complement[entry] {
                 return entry * (TARGET_SUM - entry);
             }
             has_complement[TARGET_SUM - entry] = true;
@@ -34,7 +34,7 @@ impl Solver for Day01 {
             has_complement = [false; TARGET_SUM + 1];
 
             for &entry in input.iter().filter(|&&entry| entry <= current_target) {
-                if has_complement[entry as usize] {
+                if has_complement[entry] {
                     return entry * (current_target - entry) * (TARGET_SUM - current_target);
                 }
                 has_complement[current_target - entry] = true;
@@ -52,7 +52,6 @@ impl Solver for Day01 {
             .collect::<Vec<usize>>();
 
         assert!(input.iter().all(|&entry| entry <= TARGET_SUM));
-        // eprintln!("{:?}", input);
 
         input
     }
